New function 'generateMask'
I changed the name of test.py to CropSegment.py , and I turned the program into a function that takes in the name of an image, as well as the x and y coordinates, and outputs only the segment.
